AutoRiaProject - v1.6.1 - add Premium user

+++ backend/apps/users/views.py
@@ -2,9 +2,9 @@ import os
 
 from core.permissions import IsManager, IsSuperUser, IsUser
 from core.services.email_service import EmailService
+from core.services.profanity_service import NoProfanityService
 from django.contrib.auth import get_user_model
 from django.core.mail import EmailMultiAlternatives
-from django.template.loader import get_template
 from rest_framework import status
 from rest_framework.generics import (GenericAPIView, ListCreateAPIView,
                                      UpdateAPIView)
@@ -78,16 +78,24 @@ class UsersAddCarView(GenericAPIView):
     def post(self, *args, **kwargs):
         user = self.get_object()
         user_serializer = UserSerializer(user)
-        if not user_serializer.data.get('is_premium'):
-            CarModel.objects.filter(user_id=user_serializer.data.get('id')).delete()
-            EmailService.payment(user)
-        data = self.request.data
-        data = data.copy()
-        data['user'] = user_serializer.data.get('id')
-        serializer = CarSerializer(data=data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save(user=user)
-        return Response(serializer.data, status.HTTP_201_CREATED)
+        if not user_serializer.data.get('is_active'):
+            return Response({'details': 'User is not active, please send message to admin for unban'},
+                            status=status.HTTP_400_BAD_REQUEST)
+        else:
+            data = self.request.data
+            no_profanity_data = NoProfanityService.no_profanity_check(user=user, data=data)
+            if not no_profanity_data:
+                return Response({'details': 'матюки'}, status.HTTP_400_BAD_REQUEST)
+            print(no_profanity_data)
+            data = data.copy()
+            data['user'] = user_serializer.data.get('id')
+            serializer = CarSerializer(data=data)
+            serializer.is_valid(raise_exception=True)
+            if not user_serializer.data.get('is_premium'):
+                CarModel.objects.filter(user_id=user_serializer.data.get('id')).delete()
+                EmailService.payment(user)
+            serializer.save(user=user)
+            return Response(serializer.data, status.HTTP_201_CREATED)
